backend-serverless/requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
markupsafe==2.0.1
flask-sqlalchemy==2.4.4
flask-marshmallow==0.14.0
Flask-Cors==3.0.9
python-dotenv==0.14.0